#17 [FIX] 연동 이슈 해결

+++ domain/src/main/java/com/gdg/domain/repository/CrowdZeroRepository.kt
@@ -4,6 +4,6 @@ import com.gdg.domain.entity.CongestionEntity
 import com.gdg.domain.entity.WeatherEntity
 
 interface CrowdZeroRepository {
-    suspend fun getWeather(areaId: Long): Result<WeatherEntity>
-    suspend fun getCongestion(areaId: Long): Result<CongestionEntity>
+    suspend fun getWeather(areaId: Int): Result<WeatherEntity>
+    suspend fun getCongestion(areaId: Int): Result<CongestionEntity>
 }
